requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.2
MarkupSafe==1.0
numpy==1.14.2
opencv-python==3.4.0.12
pkg-resources==0.0.0
Werkzeug==0.14.1
